split cluster parsing
