+++ Package.swift
@@ -1975,6 +1975,8 @@ let package = Package(
         "src/core/xds/grpc/xds_client_grpc.h",
         "src/core/xds/grpc/xds_cluster.cc",
         "src/core/xds/grpc/xds_cluster.h",
+        "src/core/xds/grpc/xds_cluster_parser.cc",
+        "src/core/xds/grpc/xds_cluster_parser.h",
         "src/core/xds/grpc/xds_cluster_specifier_plugin.cc",
         "src/core/xds/grpc/xds_cluster_specifier_plugin.h",
         "src/core/xds/grpc/xds_common_types.cc",
