+++ build_autogenerated.yaml
@@ -1227,6 +1227,7 @@ libs:
   - src/core/xds/grpc/xds_certificate_provider.h
   - src/core/xds/grpc/xds_client_grpc.h
   - src/core/xds/grpc/xds_cluster.h
+  - src/core/xds/grpc/xds_cluster_parser.h
   - src/core/xds/grpc/xds_cluster_specifier_plugin.h
   - src/core/xds/grpc/xds_common_types.h
   - src/core/xds/grpc/xds_common_types_parser.h
@@ -2037,6 +2038,7 @@ libs:
   - src/core/xds/grpc/xds_certificate_provider.cc
   - src/core/xds/grpc/xds_client_grpc.cc
   - src/core/xds/grpc/xds_cluster.cc
+  - src/core/xds/grpc/xds_cluster_parser.cc
   - src/core/xds/grpc/xds_cluster_specifier_plugin.cc
   - src/core/xds/grpc/xds_common_types.cc
   - src/core/xds/grpc/xds_common_types_parser.cc
