+++ config.w32
@@ -831,6 +831,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\xds\\grpc\\xds_certificate_provider.cc " +
     "src\\core\\xds\\grpc\\xds_client_grpc.cc " +
     "src\\core\\xds\\grpc\\xds_cluster.cc " +
+    "src\\core\\xds\\grpc\\xds_cluster_parser.cc " +
     "src\\core\\xds\\grpc\\xds_cluster_specifier_plugin.cc " +
     "src\\core\\xds\\grpc\\xds_common_types.cc " +
     "src\\core\\xds\\grpc\\xds_common_types_parser.cc " +
