+++ gRPC-C++.podspec
@@ -1335,6 +1335,7 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_certificate_provider.h',
                       'src/core/xds/grpc/xds_client_grpc.h',
                       'src/core/xds/grpc/xds_cluster.h',
+                      'src/core/xds/grpc/xds_cluster_parser.h',
                       'src/core/xds/grpc/xds_cluster_specifier_plugin.h',
                       'src/core/xds/grpc/xds_common_types.h',
                       'src/core/xds/grpc/xds_common_types_parser.h',
@@ -2617,6 +2618,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_certificate_provider.h',
                               'src/core/xds/grpc/xds_client_grpc.h',
                               'src/core/xds/grpc/xds_cluster.h',
+                              'src/core/xds/grpc/xds_cluster_parser.h',
                               'src/core/xds/grpc/xds_cluster_specifier_plugin.h',
                               'src/core/xds/grpc/xds_common_types.h',
                               'src/core/xds/grpc/xds_common_types_parser.h',
