+++ gRPC-Core.podspec
@@ -2090,6 +2090,8 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_client_grpc.h',
                       'src/core/xds/grpc/xds_cluster.cc',
                       'src/core/xds/grpc/xds_cluster.h',
+                      'src/core/xds/grpc/xds_cluster_parser.cc',
+                      'src/core/xds/grpc/xds_cluster_parser.h',
                       'src/core/xds/grpc/xds_cluster_specifier_plugin.cc',
                       'src/core/xds/grpc/xds_cluster_specifier_plugin.h',
                       'src/core/xds/grpc/xds_common_types.cc',
@@ -3396,6 +3398,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_certificate_provider.h',
                               'src/core/xds/grpc/xds_client_grpc.h',
                               'src/core/xds/grpc/xds_cluster.h',
+                              'src/core/xds/grpc/xds_cluster_parser.h',
                               'src/core/xds/grpc/xds_cluster_specifier_plugin.h',
                               'src/core/xds/grpc/xds_common_types.h',
                               'src/core/xds/grpc/xds_common_types_parser.h',
