+++ package.xml
@@ -1959,6 +1959,8 @@
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_client_grpc.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_cluster.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_cluster.h" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_cluster_parser.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_cluster_parser.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_cluster_specifier_plugin.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_cluster_specifier_plugin.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_common_types.cc" role="src" />
