+++ src/core/xds/grpc/xds_cluster.h
@@ -17,32 +17,17 @@
 #ifndef GRPC_SRC_CORE_XDS_GRPC_XDS_CLUSTER_H
 #define GRPC_SRC_CORE_XDS_GRPC_XDS_CLUSTER_H
 
-#include <stdint.h>
-
-#include <set>
 #include <string>
 #include <vector>
 
-#include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 #include "absl/types/variant.h"
-#include "envoy/config/cluster/v3/cluster.upbdefs.h"
-#include "envoy/extensions/clusters/aggregate/v3/cluster.upbdefs.h"
-#include "envoy/extensions/transport_sockets/tls/v3/tls.upbdefs.h"
-#include "envoy/extensions/upstreams/http/v3/http_protocol_options.upbdefs.h"
-#include "upb/reflection/def.h"
-
-#include <grpc/support/json.h>
-#include <grpc/support/port_platform.h>
 
 #include "src/core/load_balancing/outlier_detection/outlier_detection.h"
 #include "src/core/util/json/json.h"
-#include "src/core/xds/grpc/xds_bootstrap_grpc.h"
 #include "src/core/xds/grpc/xds_common_types.h"
 #include "src/core/xds/grpc/xds_health_status.h"
 #include "src/core/xds/grpc/xds_server_grpc.h"
-#include "src/core/xds/xds_client/xds_bootstrap.h"
-#include "src/core/xds/xds_client/xds_client.h"
 #include "src/core/xds/xds_client/xds_resource_type.h"
 #include "src/core/xds/xds_client/xds_resource_type_impl.h"
 
@@ -117,27 +102,6 @@ struct XdsClusterResource : public XdsResourceType::ResourceData {
   }
 
   std::string ToString() const;
-};
-
-class XdsClusterResourceType
-    : public XdsResourceTypeImpl<XdsClusterResourceType, XdsClusterResource> {
- public:
-  absl::string_view type_url() const override {
-    return "envoy.config.cluster.v3.Cluster";
-  }
-
-  DecodeResult Decode(const XdsResourceType::DecodeContext& context,
-                      absl::string_view serialized_resource) const override;
-
-  bool AllResourcesRequiredInSotW() const override { return true; }
-
-  void InitUpbSymtab(XdsClient*, upb_DefPool* symtab) const override {
-    envoy_config_cluster_v3_Cluster_getmsgdef(symtab);
-    envoy_extensions_clusters_aggregate_v3_ClusterConfig_getmsgdef(symtab);
-    envoy_extensions_transport_sockets_tls_v3_UpstreamTlsContext_getmsgdef(
-        symtab);
-    envoy_extensions_upstreams_http_v3_HttpProtocolOptions_getmsgdef(symtab);
-  }
 };
 
 }  // namespace grpc_core
