+++ src/python/grpcio/grpc_core_dependencies.py
@@ -840,6 +840,7 @@ CORE_SOURCE_FILES = [
     'src/core/xds/grpc/xds_certificate_provider.cc',
     'src/core/xds/grpc/xds_client_grpc.cc',
     'src/core/xds/grpc/xds_cluster.cc',
+    'src/core/xds/grpc/xds_cluster_parser.cc',
     'src/core/xds/grpc/xds_cluster_specifier_plugin.cc',
     'src/core/xds/grpc/xds_common_types.cc',
     'src/core/xds/grpc/xds_common_types_parser.cc',
