+++ tools/doxygen/Doxyfile.core.internal
@@ -2761,6 +2761,8 @@ src/core/xds/grpc/xds_client_grpc.cc \
 src/core/xds/grpc/xds_client_grpc.h \
 src/core/xds/grpc/xds_cluster.cc \
 src/core/xds/grpc/xds_cluster.h \
+src/core/xds/grpc/xds_cluster_parser.cc \
+src/core/xds/grpc/xds_cluster_parser.h \
 src/core/xds/grpc/xds_cluster_specifier_plugin.cc \
 src/core/xds/grpc/xds_cluster_specifier_plugin.h \
 src/core/xds/grpc/xds_common_types.cc \
